+++ docs/changelog.rst
@@ -11,6 +11,11 @@ Beets now requires Python 3.10 or later since support for EOL Python 3.9 has
 been dropped.
 
 New features:
+* :doc:`plugins/smartplaylist`: Add new configuration option ``dest_regen`` to
+  regenerate items' path in the generated playlist instead of using the ones of
+  the library. This is useful when items have been imported in don't copy-move
+  (`-C -M`) mode in the library but are later passed through the `convert`
+  plugin which will regenerate new paths according to the Beets path format.
 
 - :doc:`plugins/ftintitle`: Added argument for custom feat. words in ftintitle.
 - :doc:`plugins/ftintitle`: Added album template value ``album_artist_no_feat``.